refactor: split test_booking into test_models and test_views
