+++ booking/tests/test_views.py
@@ -1,68 +1,15 @@
-from django.contrib.auth import get_user_model
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient
 
-from booking.models import Order, get_deleted_user, Ticket
+from booking.models import Order
 from booking.serializers import OrderListSerializer, OrderSerializer
-from fleet.tests.test_fleet import sample_airplane_type, sample_airplane
-from flight_ops.tests.test_flight_ops import sample_flight, sample_route
-from location.tests.test_location import (
-    sample_country, sample_city, sample_airport
-)
+from flight_ops.tests.utils import sample_flight
+from user.tests.utils import sample_user
 
 ORDER_URL = reverse("booking:order-list")
-
-
-def sample_user(**params):
-    defaults = {
-        "email": "test_user@test.com",
-        "password": "test1234",
-    }
-    defaults.update(params)
-
-    return get_user_model().objects.create_user(**defaults)
-
-
-class OrderModelTest(TestCase):
-    def setUp(self):
-        self.user = sample_user()
-        self.order = Order.objects.create(user=self.user)
-
-    def test_order_str(self):
-        self.assertEqual(str(self.order), f"order {self.order.id}")
-
-    def test_deleted_user(self):
-        deleted_user = get_deleted_user()
-        self.assertEqual(deleted_user.email, "deleted_user@airport.com")
-
-        self.user.delete()
-        self.order.refresh_from_db()
-        self.assertEqual(self.order.user, deleted_user)
-
-
-class TicketModelTest(TestCase):
-    def setUp(self):
-        country = sample_country()
-        city = sample_city(country=country)
-        source = sample_airport(closest_big_city=city)
-        destination = sample_airport(name="Gatwick", closest_big_city=city)
-        route = sample_route(source=source, destination=destination)
-        airplane_type = sample_airplane_type()
-        airplane = sample_airplane(airplane_type=airplane_type)
-        self.flight = sample_flight(route=route, airplane=airplane)
-
-    def test_ticket_str(self):
-        order = Order.objects.create(user=sample_user())
-        ticket = Ticket.objects.create(
-            row=1, seat=5, flight=self.flight, order=order
-        )
-        self.assertEqual(
-            f"Ticket for row 1, seat 5, flight {self.flight.number}",
-            str(ticket)
-        )
 
 
 class UnauthenticatedOrderAPITests(TestCase):
@@ -80,14 +27,7 @@ class AuthenticatedOrderAPITests(TestCase):
         self.user = sample_user()
         self.client.force_authenticate(user=self.user)
 
-        country = sample_country()
-        city = sample_city(country=country)
-        source = sample_airport(closest_big_city=city)
-        destination = sample_airport(name="Gatwick", closest_big_city=city)
-        route = sample_route(source=source, destination=destination)
-        airplane_type = sample_airplane_type()
-        airplane = sample_airplane(airplane_type=airplane_type)
-        self.flight = sample_flight(route=route, airplane=airplane)
+        self.flight = sample_flight()
 
     def test_orders_list_has_auth_user_orders(self):
         test_user_order = Order.objects.create(user=self.user)
